pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gustavo Rudolf", email = "gustavorudolfifc@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

